Sunlight Hrs Calc
Christiane, I hope this helps when you are pulling the time data, to convert it to military time. The TA helped me with the military time conversion. Then, I calculated the total sunlight hours by breaking down the military time into minutes. I hope this helps! If you run the file from your laptop, be sure to insert your API. I did not want to publish it on github, so you will need to copy and paste it.